added auth-utility to get rid of some linting errors
