+++ src/app/game/[id]/page.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable */
 "use client";
 
 import { useEffect, useState } from 'react';
