+++ src/app/games/page.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable */
 "use client";
 
 import Link from 'next/link';
