+++ src/components/GameCreator.tsx
@@ -1,3 +1,5 @@
+/* eslint-disable */
+
 "use client";
 
 import { useState } from "react";
